Added support for auth

+++ src/routes/Group.controller.ts
@@ -7,20 +7,25 @@ import { GroupModel, Group } from "../models/Group.model";
 import { DatabaseService } from "../services/database/database.service";
 
 import { ResponseDto } from "../dto/response.dto";
+import { DefaultProcessedAuthHeader, authMiddleware } from "../middleware/auth";
+import RoleService from "../services/role/Role.service";
+import { ROLE } from "../constants/Role.dictionary";
 
 /*-------- Initialization--------*/
 const router = express.Router();
 const db = new DatabaseService();
+const rs = new RoleService();
 
 /*-------- Dto --------*/
 
 interface GetGroup {
   headers: {
     groupid: string;
+    email: string;
   };
 }
 
-interface CreateGroup {
+interface CreateGroup extends DefaultProcessedAuthHeader {
   body: Group;
 }
 
@@ -32,10 +37,14 @@ interface UpdateGroup extends GetGroup {
 /*-------- Methods --------*/
 
 const getGroup = async (request: GetGroup): Promise<ResponseDto> => {
+  if (!(await rs.verifyRole(request.headers.email, ROLE.GROUP_READ))) {
+    return { error: "User does not exist or doesn't have this role" };
+  }
   try {
     const group = await db.findOne(GroupModel, {
       _id: request.headers.groupid,
     });
+
     return {
       data: group,
     };
@@ -45,6 +54,9 @@ const getGroup = async (request: GetGroup): Promise<ResponseDto> => {
 };
 
 const createGroup = async (request: CreateGroup): Promise<ResponseDto> => {
+  if (!(await rs.verifyRole(request.headers.email, ROLE.GROUP_CREATE))) {
+    return { error: "User does not exist or doesn't have this role" };
+  }
   try {
     const group = await db.create(GroupModel, request.body);
     return {
@@ -56,6 +68,9 @@ const createGroup = async (request: CreateGroup): Promise<ResponseDto> => {
 };
 
 const updateGroup = async (request: UpdateGroup): Promise<ResponseDto> => {
+  if (!(await rs.verifyRole(request.headers.email, ROLE.GROUP_UPDATE))) {
+    return { error: "User does not exist or doesn't have this role" };
+  }
   try {
     const group = await db.findOneAndUpdate(
       GroupModel,
@@ -74,6 +89,9 @@ const updateGroup = async (request: UpdateGroup): Promise<ResponseDto> => {
 };
 
 const deleteGroup = async (request: GetGroup): Promise<ResponseDto> => {
+  if (!(await rs.verifyRole(request.headers.email, ROLE.GROUP_DELETE))) {
+    return { error: "User does not exist or doesn't have this role" };
+  }
   try {
     await GroupModel.findOneAndRemove({
       _id: request.headers.groupid,
@@ -88,23 +106,39 @@ const deleteGroup = async (request: GetGroup): Promise<ResponseDto> => {
 
 /*-------- Routes --------*/
 
-router.get("/id", async (request: GetGroup, response: any) => {
-  const group: ResponseDto = await getGroup(request);
-  response.send(group);
-});
+router.get(
+  "/id",
+  [authMiddleware],
+  async (request: GetGroup, response: any) => {
+    const group: ResponseDto = await getGroup(request);
+    response.send(group);
+  }
+);
 
-router.post("/", async (request: CreateGroup, response: any) => {
-  const group: ResponseDto = await createGroup(request);
-  response.send(group);
-});
+router.post(
+  "/",
+  [authMiddleware],
+  async (request: CreateGroup, response: any) => {
+    const group: ResponseDto = await createGroup(request);
+    response.send(group);
+  }
+);
 
-router.put("/id", async (request: UpdateGroup, response: any) => {
-  const group: ResponseDto = await updateGroup(request);
-  response.send(group);
-});
+router.put(
+  "/id",
+  [authMiddleware],
+  async (request: UpdateGroup, response: any) => {
+    const group: ResponseDto = await updateGroup(request);
+    response.send(group);
+  }
+);
 
-router.delete("/id", async (request: UpdateGroup, response: any) => {
-  const group: ResponseDto = await deleteGroup(request);
-  response.send(group);
-});
+router.delete(
+  "/id",
+  [authMiddleware],
+  async (request: UpdateGroup, response: any) => {
+    const group: ResponseDto = await deleteGroup(request);
+    response.send(group);
+  }
+);
 export default router;
